fix(pdf): restore stable local PDF.js viewer
