--- .github/workflows/apply-pdfjs-stable-viewer.yml
@@ -1,50 +1,0 @@
-name: Apply PDF.js stability correction
-
-'on':
-  push:
-    branches:
-      - fix/pdfjs-stable-viewer
-    paths:
-      - .github/workflows/apply-pdfjs-stable-viewer.yml
-
-permissions:
-  contents: write
-
-jobs:
-  apply-fix:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Check out correction branch
-        uses: actions/checkout@v4
-        with:
-          ref: fix/pdfjs-stable-viewer
-          fetch-depth: 0
-
-      - name: Apply correction
-        run: python3 tools/apply_pdfjs_stable_viewer.py
-
-      - name: Remove one-time automation files
-        run: rm .github/workflows/apply-pdfjs-stable-viewer.yml tools/apply_pdfjs_stable_viewer.py
-
-      - name: Validate changed files
-        run: |
-          php -l view.php
-          php -l mod_form.php
-          php -l locallib.php
-          php -l db/upgrade.php
-          php -l version.php
-          php -l tests/pdf_displaymode_test.php
-          python3 -c "import xml.etree.ElementTree as ET; ET.parse('db/install.xml'); ET.parse('thirdpartylibs.xml')"
-          grep -q "mod_videoplayer/pdfviewer" view.php
-          test "$(grep -c "mod_videoplayer/ebookviewer" view.php)" -eq 0
-          test "$(grep -c "thirdpartylibs/pageflip" view.php)" -eq 0
-          grep -q 'DEFAULT="pdfjs"' db/install.xml
-          grep -q '2026080600' version.php
-
-      - name: Commit correction
-        run: |
-          git config user.name "Elearning Cloud"
-          git config user.email "94720805+JoseMorenoSalgado@users.noreply.github.com"
-          git add -A
-          git commit -m "fix(pdf): restore stable local PDF.js viewer"
-          git push origin HEAD:fix/pdfjs-stable-viewer
