+++ db/install.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" ?>
-<XMLDB PATH="mod/videoplayer/db" VERSION="20260805" COMMENT="XMLDB file for Moodle mod_videoplayer"
+<XMLDB PATH="mod/videoplayer/db" VERSION="20260806" COMMENT="XMLDB file for Moodle mod_videoplayer"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:noNamespaceSchemaLocation="../../../lib/xmldb/xmldb.xsd"
 >
@@ -17,7 +17,7 @@
         <FIELD NAME="source" TYPE="char" LENGTH="32" NOTNULL="true" DEFAULT="googledrive"/>
         <FIELD NAME="videourl" TYPE="char" LENGTH="1024" NOTNULL="false"/>
         <FIELD NAME="type" TYPE="char" LENGTH="32" NOTNULL="true" DEFAULT="auto"/>
-        <FIELD NAME="displaymode" TYPE="char" LENGTH="32" NOTNULL="true" DEFAULT="standard"/>
+        <FIELD NAME="displaymode" TYPE="char" LENGTH="32" NOTNULL="true" DEFAULT="pdfjs"/>
         <FIELD NAME="disabledownload" TYPE="int" LENGTH="1" NOTNULL="true" DEFAULT="1"/>
         <FIELD NAME="disablecontextmenu" TYPE="int" LENGTH="1" NOTNULL="true" DEFAULT="1"/>
         <FIELD NAME="enablewatermark" TYPE="int" LENGTH="1" NOTNULL="true" DEFAULT="0"/>
