--- tools/apply_pdfjs_stable_viewer.py
@@ -1,337 +1,0 @@
-from pathlib import Path
-import re
-
-
-def replace_once(path: str, old: str, new: str) -> None:
-    file = Path(path)
-    content = file.read_text(encoding='utf-8')
-    if old not in content:
-        raise SystemExit(f'Expected block not found in {path}')
-    file.write_text(content.replace(old, new, 1), encoding='utf-8')
-
-
-def regex_once(path: str, pattern: str, replacement: str) -> None:
-    file = Path(path)
-    content = file.read_text(encoding='utf-8')
-    updated, count = re.subn(pattern, replacement, content, count=1, flags=re.S)
-    if count != 1:
-        raise SystemExit(f'Expected one regex match in {path}, found {count}')
-    file.write_text(updated, encoding='utf-8')
-
-
-def append_section(path: str, marker: str, section: str) -> None:
-    file = Path(path)
-    content = file.read_text(encoding='utf-8').rstrip()
-    if marker in content:
-        return
-    file.write_text(content + '\n\n' + section.strip() + '\n', encoding='utf-8')
-
-
-replace_once(
-    'view.php',
-    "require_once(__DIR__ . '/lib.php');\n",
-    "require_once(__DIR__ . '/lib.php');\nrequire_once(__DIR__ . '/locallib.php');\n"
-)
-
-regex_once(
-    'view.php',
-    r"\$displaymode = clean_param\(\$videoplayer->displaymode \?\? 'ebook', PARAM_ALPHANUMEXT\);\n"
-    r"if \(\$displaymode === 'standard'\) \{\n"
-    r"    \$displaymode = 'ebook';\n"
-    r"\}\n"
-    r"if \(!in_array\(\$displaymode, \['ebook', 'pdfjs', 'book'\], true\)\) \{\n"
-    r"    \$displaymode = 'ebook';\n"
-    r"\}",
-    "$displaymode = videoplayer_get_safe_pdf_displaymode($videoplayer->displaymode ?? null);"
-)
-
-regex_once(
-    'view.php',
-    r"if \(\$type === 'pdf'\) \{\n"
-    r"    if \(\$displaymode === 'ebook'\) \{.*?"
-    r"\n\} else if \(\$type === 'video'\) \{",
-    "if ($type === 'pdf') {\n"
-    "    $PAGE->requires->js_call_amd('mod_videoplayer/pdfviewer', 'init');\n"
-    "} else if ($type === 'video') {"
-)
-
-replace_once(
-    'view.php',
-    "    'ebookmode' => $displaymode === 'ebook',\n    'bookmode' => $displaymode === 'book',\n",
-    "    'ebookmode' => false,\n    'bookmode' => false,\n"
-)
-
-regex_once(
-    'view.php',
-    r"if \(\$type === 'pdf'\) \{\n"
-    r"    if \(\$displaymode === 'book'\) \{.*?"
-    r"\n\} else if \(\$type === 'video'\) \{",
-    "if ($type === 'pdf') {\n"
-    "    echo $OUTPUT->render_from_template('mod_videoplayer/pdfjs', $templatecontext);\n"
-    "} else if ($type === 'video') {"
-)
-
-replace_once(
-    'mod_form.php',
-    "require_once($CFG->dirroot . '/course/moodleform_mod.php');\n",
-    "require_once($CFG->dirroot . '/course/moodleform_mod.php');\nrequire_once(__DIR__ . '/locallib.php');\n"
-)
-
-regex_once(
-    'mod_form.php',
-    r"        \$displaymodes = \[.*?"
-    r"        \$mform->addHelpButton\('displaymode', 'displaymode', 'mod_videoplayer'\);\n",
-    "        $mform->addElement('hidden', 'displaymode', 'pdfjs');\n"
-    "        $mform->setType('displaymode', PARAM_ALPHANUMEXT);\n"
-)
-
-regex_once(
-    'mod_form.php',
-    r"        \$displaymode = clean_param\(\$defaultvalues\['displaymode'\] \?\? 'ebook', PARAM_ALPHANUMEXT\);\n"
-    r"        if \(\$displaymode === 'standard'\) \{\n"
-    r"            \$displaymode = 'ebook';\n"
-    r"        \}\n"
-    r"        if \(!in_array\(\$displaymode, \['ebook', 'pdfjs'\], true\)\) \{\n"
-    r"            \$displaymode = 'ebook';\n"
-    r"        \}\n"
-    r"        \$defaultvalues\['displaymode'\] = \$displaymode;",
-    "        $defaultvalues['displaymode'] = videoplayer_get_safe_pdf_displaymode(\n"
-    "            $defaultvalues['displaymode'] ?? null\n"
-    "        );"
-)
-
-locallib = Path('locallib.php')
-locallib_content = locallib.read_text(encoding='utf-8').rstrip()
-if 'function videoplayer_get_safe_pdf_displaymode' not in locallib_content:
-    locallib_content += """
-
-/**
- * Return the production-safe PDF display mode.
- *
- * The animated PageFlip and legacy book renderers remain bundled for future
- * compatibility work, but they are not executed in the stable release path.
- * This prevents JavaScript assets from being exposed as document content and
- * keeps mobile rendering on the locally bundled PDF.js implementation.
- *
- * @param string|null $requestedmode Requested or legacy database value.
- * @return string Always returns the supported PDF.js mode.
- */
-function videoplayer_get_safe_pdf_displaymode(?string $requestedmode): string {
-    $requestedmode = clean_param((string) $requestedmode, PARAM_ALPHANUMEXT);
-    $supportedmodes = ['pdfjs'];
-
-    return in_array($requestedmode, $supportedmodes, true) ? $requestedmode : 'pdfjs';
-}
-"""
-    locallib.write_text(locallib_content + '\n', encoding='utf-8')
-
-replace_once('db/install.xml', 'VERSION="20260805"', 'VERSION="20260806"')
-replace_once(
-    'db/install.xml',
-    '<FIELD NAME="displaymode" TYPE="char" LENGTH="32" NOTNULL="true" DEFAULT="standard"/>',
-    '<FIELD NAME="displaymode" TYPE="char" LENGTH="32" NOTNULL="true" DEFAULT="pdfjs"/>'
-)
-
-upgrade = Path('db/upgrade.php')
-upgrade_content = upgrade.read_text(encoding='utf-8')
-if "if ($oldversion < 2026080600)" not in upgrade_content:
-    step = """
-
-    if ($oldversion < 2026080600) {
-        $table = new xmldb_table('videoplayer');
-        $displaymodefield = new xmldb_field(
-            'displaymode',
-            XMLDB_TYPE_CHAR,
-            '32',
-            null,
-            XMLDB_NOTNULL,
-            null,
-            'pdfjs',
-            'type'
-        );
-
-        if ($dbman->table_exists($table) && $dbman->field_exists($table, $displaymodefield)) {
-            $dbman->change_field_default($table, $displaymodefield);
-            $DB->set_field_select(
-                'videoplayer',
-                'displaymode',
-                'pdfjs',
-                'displaymode IS NULL OR displaymode <> :displaymode',
-                ['displaymode' => 'pdfjs']
-            );
-        }
-
-        upgrade_mod_savepoint(true, 2026080600, 'videoplayer');
-    }
-"""
-    needle = "\n    return true;\n}"
-    if needle not in upgrade_content:
-        raise SystemExit('Upgrade return block not found')
-    upgrade.write_text(upgrade_content.replace(needle, step + needle, 1), encoding='utf-8')
-
-replace_once('version.php', '$plugin->version = 2026080505;', '$plugin->version = 2026080600;')
-replace_once('version.php', "$plugin->release = '1.1.26-beta';", "$plugin->release = '1.1.27-beta';")
-
-Path('tests/pdf_displaymode_test.php').write_text("""<?php
-// This file is part of Moodle - http://moodle.org/
-//
-// Moodle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-namespace mod_videoplayer;
-
-use PHPUnit\\Framework\\Attributes\\CoversNothing;
-use PHPUnit\\Framework\\Attributes\\DataProvider;
-
-/**
- * PDF display-mode stability tests.
- *
- * @package    mod_videoplayer
- * @category   test
- * @copyright  2026 Jose Erasmo Moreno Salgado - Elearning Cloud
- * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-#[CoversNothing]
-final class pdf_displaymode_test extends \\advanced_testcase {
-    /**
-     * Legacy and unsupported modes must resolve to the stable PDF.js viewer.
-     *
-     * @param string|null $requestedmode Requested mode.
-     */
-    #[DataProvider('legacy_mode_provider')]
-    public function test_legacy_modes_resolve_to_pdfjs(?string $requestedmode): void {
-        require_once(__DIR__ . '/../locallib.php');
-
-        $this->assertSame('pdfjs', videoplayer_get_safe_pdf_displaymode($requestedmode));
-    }
-
-    /**
-     * Supply current, legacy and malformed display-mode values.
-     *
-     * @return array<string, array{0: string|null}>
-     */
-    public static function legacy_mode_provider(): array {
-        return [
-            'current PDF.js mode' => ['pdfjs'],
-            'legacy standard mode' => ['standard'],
-            'PageFlip ebook mode' => ['ebook'],
-            'legacy book mode' => ['book'],
-            'empty mode' => [''],
-            'missing mode' => [null],
-            'invalid mode' => ['../../pageflip'],
-        ];
-    }
-
-    /**
-     * The learner entry point must not load alternative PDF renderers.
-     */
-    public function test_view_uses_only_pdfjs_renderer(): void {
-        $viewsource = file_get_contents(__DIR__ . '/../view.php');
-
-        $this->assertIsString($viewsource);
-        $this->assertStringContainsString("mod_videoplayer/pdfviewer", $viewsource);
-        $this->assertStringNotContainsString("mod_videoplayer/ebookviewer", $viewsource);
-        $this->assertStringNotContainsString("mod_videoplayer/bookviewer", $viewsource);
-        $this->assertStringNotContainsString("thirdpartylibs/pageflip", $viewsource);
-    }
-}
-""", encoding='utf-8')
-
-append_section(
-    'README.md',
-    '## PDF viewer stability policy (1.1.27-beta)',
-    """
-## PDF viewer stability policy (1.1.27-beta)
-
-Drive Resource now routes every PDF through the locally bundled PDF.js viewer. Legacy `standard`, `ebook` and `book` database values are normalised to `pdfjs` during upgrade and at runtime. StPageFlip remains vendored only for future compatibility work and is not loaded in the production learner path. This removes the mobile failure mode where a JavaScript asset could be exposed as visible document content.
-
-After deploying this release, run the Moodle upgrade, purge Moodle caches and invalidate any reverse-proxy or CDN cache that may still contain older viewer assets.
-"""
-)
-
-changelog = Path('CHANGELOG.md')
-changelog_content = changelog.read_text(encoding='utf-8')
-if '## 1.1.27-beta - 2026-08-06' not in changelog_content:
-    release = """
-## 1.1.27-beta - 2026-08-06
-
-### Fixed
-
-- Restored the local PDF.js viewer as the only production PDF rendering path.
-- Prevented StPageFlip and legacy book assets from loading in learner sessions.
-- Normalised legacy `standard`, `ebook` and `book` modes to `pdfjs`.
-- Added upgrade handling, regression tests and deployment verification guidance.
-
-"""
-    first_break = changelog_content.find('\n', changelog_content.find('#'))
-    if first_break == -1:
-        changelog_content = changelog_content.rstrip() + '\n\n' + release
-    else:
-        changelog_content = changelog_content[:first_break + 1] + '\n' + release + changelog_content[first_break + 1:]
-    changelog.write_text(changelog_content, encoding='utf-8')
-
-append_section(
-    'docs/architecture.md',
-    '## Stable PDF rendering boundary',
-    """
-## Stable PDF rendering boundary
-
-The production PDF pipeline is now strictly `protected.php -> authenticated range proxy -> local PDF.js -> canvas`. The learner-facing entry point does not load StPageFlip or the legacy book renderer. Historical display-mode values are converted to `pdfjs`, keeping one deterministic rendering boundary across desktop and mobile devices.
-"""
-)
-
-append_section(
-    'docs/developer-guide.md',
-    '## PDF renderer contract',
-    """
-## PDF renderer contract
-
-All production PDF changes must preserve `videoplayer_get_safe_pdf_displaymode()` and the single PDF.js rendering path. Do not reintroduce PageFlip from `view.php`. Any future animated reader must be isolated behind an explicit experimental feature flag, include a mobile PDF.js fallback, remain idempotent, and pass physical-device regression tests before release.
-"""
-)
-
-append_section(
-    'docs/security.md',
-    '## PDF asset isolation',
-    """
-## PDF asset isolation
-
-Only the protected resource URL may be supplied to PDF.js. Third-party JavaScript paths must never be accepted as document URLs or rendered inside content containers. The stable viewer does not execute StPageFlip, reducing the attack surface and preventing library source from being exposed as learner-visible content. Server-side login, enrolment and capability checks remain mandatory in `protected.php`.
-"""
-)
-
-append_section(
-    'docs/installation.md',
-    '## Upgrade to 1.1.27-beta',
-    """
-## Upgrade to 1.1.27-beta
-
-1. Replace the plugin code and visit Site administration > Notifications, or run the Moodle CLI upgrade.
-2. Confirm the `videoplayer.displaymode` default is `pdfjs`; the upgrade converts legacy modes automatically.
-3. Purge all Moodle caches and reset PHP OPcache.
-4. Purge Cloudflare, NGINX or other reverse-proxy caches for `/mod/videoplayer/`.
-5. Test one protected PDF on Android, iPhone, tablet and desktop. The response from `protected.php` must start with `%PDF-` and use `Content-Type: application/pdf`.
-"""
-)
-
-append_section(
-    'docs/manual-test-checklist.md',
-    '## PDF.js stability regression',
-    """
-## PDF.js stability regression
-
-- Create or edit a PDF resource that previously stored `ebook`, `book` or `standard`; verify it opens with PDF.js.
-- Confirm no request for `thirdpartylibs/pageflip/page-flip.browser.js` occurs in the browser network panel.
-- Verify previous/next page, zoom, fit, fullscreen, swipe navigation and progress persistence.
-- Verify `protected.php` returns `%PDF-`, `application/pdf`, `Accept-Ranges: bytes` and valid partial responses.
-- Repeat on a physical iPhone and Android phone after clearing browser and Moodle caches.
-"""
-)
-
-replace_once(
-    'thirdpartylibs.xml',
-    '<description>Optional local page turning library used by Drive Resource protected ebook mode.</description>',
-    '<description>Vendored compatibility library. It is not loaded by the stable production PDF.js viewer.</description>'
-)
