+++ version.php
@@ -26,8 +26,8 @@ defined('MOODLE_INTERNAL') || die();
 
 $plugin = new stdClass();
 $plugin->component = 'mod_videoplayer';
-$plugin->version = 2026080505;
-$plugin->release = '1.1.26-beta';
+$plugin->version = 2026080600;
+$plugin->release = '1.1.27-beta';
 $plugin->requires = 2025041400;
 $plugin->supported = [500, 502];
 $plugin->maturity = MATURITY_BETA;
